--- frontend/__tests__/components/features/sidebar/user-actions.test.tsx
@@ -1,110 +1,0 @@
-import { screen, fireEvent, act } from "@testing-library/react";
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { renderWithProviders } from "test-utils";
-import { createRoutesStub } from "react-router";
-import { UserActions } from "#/components/features/sidebar/user-actions";
-import { useSelectedOrganizationStore } from "#/stores/selected-organization-store";
-import { server } from "#/mocks/node";
-import { http, HttpResponse } from "msw";
-import { createMockWebClientConfig } from "#/mocks/settings-handlers";
-
-const RouterStub = createRoutesStub([
-  {
-    path: "/",
-    Component: () => (
-      <UserActions user={{ avatar_url: "https://example.com/avatar.png" }} />
-    ),
-  },
-]);
-
-const renderUserActions = () => {
-  return renderWithProviders(<RouterStub initialEntries={["/"]} />);
-};
-
-describe("UserActions", () => {
-  beforeEach(() => {
-    useSelectedOrganizationStore.setState({ organizationId: "1" });
-
-    // Mock config to return SaaS mode so useShouldShowUserFeatures returns true
-    server.use(
-      http.get("/api/v1/web-client/config", () =>
-        HttpResponse.json(createMockWebClientConfig({ app_mode: "saas" })),
-      ),
-    );
-  });
-
-  afterEach(() => {
-    vi.clearAllMocks();
-    server.resetHandlers();
-  });
-
-  describe("menu close delay", () => {
-    beforeEach(() => {
-      vi.useFakeTimers();
-    });
-
-    afterEach(() => {
-      vi.useRealTimers();
-    });
-
-    it("should keep menu visible when mouse leaves and re-enters within 500ms", async () => {
-      // Arrange - render and wait for queries to settle
-      renderUserActions();
-      await act(async () => {
-        await vi.runAllTimersAsync();
-      });
-
-      const userActions = screen.getByTestId("user-actions");
-
-      // Act - open menu
-      await act(async () => {
-        fireEvent.mouseEnter(userActions);
-      });
-
-      // Assert - menu is visible
-      expect(screen.getByTestId("user-context-menu")).toBeInTheDocument();
-
-      // Act - leave and re-enter within 500ms
-      await act(async () => {
-        fireEvent.mouseLeave(userActions);
-        await vi.advanceTimersByTimeAsync(200);
-        fireEvent.mouseEnter(userActions);
-      });
-
-      // Assert - menu should still be visible after waiting (pending close was cancelled)
-      await act(async () => {
-        await vi.advanceTimersByTimeAsync(500);
-      });
-      expect(screen.getByTestId("user-context-menu")).toBeInTheDocument();
-    });
-
-    it("should not close menu before 500ms delay when mouse leaves", async () => {
-      // Arrange - render and wait for queries to settle
-      renderUserActions();
-      await act(async () => {
-        await vi.runAllTimersAsync();
-      });
-
-      const userActions = screen.getByTestId("user-actions");
-
-      // Act - open menu
-      await act(async () => {
-        fireEvent.mouseEnter(userActions);
-      });
-
-      // Assert - menu is visible
-      expect(screen.getByTestId("user-context-menu")).toBeInTheDocument();
-
-      // Act - leave without re-entering, but check before timeout expires
-      await act(async () => {
-        fireEvent.mouseLeave(userActions);
-        await vi.advanceTimersByTimeAsync(400); // Before the 500ms delay
-      });
-
-      // Assert - menu should still be visible (delay hasn't expired yet)
-      // Note: The menu is always in DOM but with opacity-0 when closed.
-      // This test verifies the state hasn't changed yet (delay is working).
-      expect(screen.getByTestId("user-context-menu")).toBeInTheDocument();
-    });
-  });
-});
